fixed texts in miscs (moved to directory)
