+++ src/misc/texts/loginFormLinks.js
