+++ src/misc/texts/registerPageText.js
